Rename: 컴포넌트에 스켈레톤 디렉토리 분리
